Add files via upload
This is a course intended for Computational Biology seniors who want to synthesize all of the learning that they have done for their degree and add it to professional behavioural expectations. It is a review of technical skills (these are skills already taken in courses required for their major) and adds in 'soft' or professional skills.